+++ src/pipette/__init__.py
@@ -1,63 +1,30 @@
-"""Pipette: Unified import and export of data in Python."""
+"""Pipette: Unified reading and writing of data in Python."""
 
-from pipette._atomize import atomize
 from pipette._cache_url import cache_url
-from pipette._checksums import md5, sha256
-from pipette._decode import decode
-from pipette._droplevels import droplevels
-from pipette._encode import encode
-from pipette._export import export_data
-from pipette._factorize import factorize
-from pipette._file_utils import init_dir
+from pipette._categorize import categorize, uncategorize
+from pipette._drop_nested_columns import drop_nested_columns
 from pipette._fill_lines import fill_lines
-from pipette._get_json import get_json
-from pipette._get_url_dir_list import get_url_dir_list
-from pipette._globals import NA_STRINGS, PIPETTE_TESTS_URL
-from pipette._import import import_data
-from pipette._load_data import load_data
-from pipette._load_remote_data import load_remote_data
-from pipette._match_rowname_column import match_rowname_column
-from pipette._metadata2 import metadata2
-from pipette._paste_url import paste_url
-from pipette._remove_na import remove_na
+from pipette._globals import NA_STRINGS
+from pipette._join_url import join_url
+from pipette._list_remote_dir import list_remote_dir
+from pipette._read import read
 from pipette._sanitize_na import sanitize_na
 from pipette._sanitize_percent import sanitize_percent
-from pipette._save_data import save_data
-from pipette._save_load import assign_and_save_data, load_data_as_name
 from pipette._transmit import transmit
-from pipette._unfactorize import unfactorize
-
-na_strings = NA_STRINGS
+from pipette._write import write
 
 __all__ = [
     "NA_STRINGS",
-    "PIPETTE_TESTS_URL",
-    "assign_and_save_data",
-    "atomize",
     "cache_url",
-    "decode",
-    "droplevels",
-    "encode",
-    "export_data",
-    "factorize",
+    "categorize",
+    "drop_nested_columns",
     "fill_lines",
-    "get_json",
-    "get_url_dir_list",
-    "import_data",
-    "init_dir",
-    "load_data",
-    "load_data_as_name",
-    "load_remote_data",
-    "match_rowname_column",
-    "md5",
-    "metadata2",
-    "na_strings",
-    "paste_url",
-    "remove_na",
+    "join_url",
+    "list_remote_dir",
+    "read",
     "sanitize_na",
     "sanitize_percent",
-    "save_data",
-    "sha256",
     "transmit",
-    "unfactorize",
+    "uncategorize",
+    "write",
 ]